# swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Book Marketplace API"
  description: "API for managing books in the Book Marketplace application"
basePath: "/api"
...
definitions:
  Book:
    type: "object"
    properties:
      bookId:
        type: "integer"
      ISBN:
        type: "string"
      title:
        type: "string"
      author:
        type: "string"
      edition:
        type: "integer"
      publicationYear:
        type: "integer"
      language:
        type: "string"
      publisher:
        type: "string"
      originalPrice:
        type: "number"
      currentPrice:
        type: "number"
      description:
        type: "string"
      category:
        type: "string"
      conditionDescription:
        type: "string"
      condition:
        type: "string"
      status:
        type: "string"
  User:
    type: "object"
    description: "Represents a user within the marketplace."
    properties:
      userId:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      phoneNumber:
        type: "string"
      funds:
        type: "number"
        format: "double"
      transactions:
        type: "array"
        items:
          $ref: "#/definitions/Transaction"
  Transaction:
    type: "object"
    description: "Represents a transaction made within the marketplace."
    properties:
      transactionId:
        type: "integer"
      user:
        $ref: "#/definitions/User"
      book:
        $ref: "#/definitions/Book"
      transactionDate:
        type: "string"
        format: "date-time"
      transactionAmount:
        type: "number"
      status:
        type: "string"
        enum:
          - "PENDING"
          - "COMPLETED"
          - "CANCELLED"
      transactionNotes:
        type: "string"
      transactionType:
        type: "string"
        enum:
          - "BUY"
          - "SELL"
...
paths:
...
  /users/:
    get:
      tags:
        - "Users"
      description: "Retrieve all users."
      responses:
        200:
          description: "List of users."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        204:
          description: "No users found."
        500:
          description: "Internal Server Error."
  /users/getUserById:
    get:
      tags:
        - "Users"
      description: "Retrieve a user by their ID."
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
      responses:
        200:
          description: "User details."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid user ID."
        404:
          description: "User not found."
        500:
          description: "Internal Server Error."
  /users/searchUsers:
    get:
      tags:
        - "Users"
      description: "Search for a user by username or email."
      parameters:
        - name: keyword
          in: query
          type: string
          description: "Search term for username or email."
      responses:
        200:
          description: "User details."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid search term."
        404:
          description: "User not found."
        500:
          description: "Internal Server Error."
  /users/addUser:
    post:
      tags:
        - "Users"
      description: "Add a new user."
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user
          in: body
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "Newly added user."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Validation error."
        409:
          description: "User already exists."
        500:
          description: "Internal Server Error."
  /users/updateUser:
    put:
      tags:
        - "Users"
      description: "Update an existing user."
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
        - name: user
          in: body
          schema:
            $ref: "#/definitions/User"
      responses:
        200:
          description: "Updated user details."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid user ID or validation error."
        404:
          description: "User not found."
        500:
          description: "Internal Server Error."
  /users/deleteUser:
    delete:
      tags:
        - "Users"
      description: "Delete a user by their ID."
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
      responses:
        200:
          description: "User deleted successfully."
        400:
          description: "Invalid user ID."
        404:
          description: "User not found."
        500:
          description: "Internal Server Error."
  /users/purchasedBooks:
    get:
      tags:
        - "Users"
      description: "Retrieve books purchased by a user."
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
      responses:
        200:
          description: "List of purchased books."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        400:
          description: "Invalid user ID."
        404:
          description: "No purchased books found."
        500:
          description: "Internal Server Error."
  /users/booksSoldByUser:
    get:
      tags:
        - "Users"
      description: "Retrieve books sold by a user."
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
      responses:
        200:
          description: "List of books sold."
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Book"
        400:
          description: "Invalid user ID."
        404:
          description: "No books sold by the user."
        500:
          description: "Internal Server Error."
  /users/addFunds:
    put:
      tags:
        - "Users"
      description: "Add funds to a user's account."
      parameters:
        - name: userId
          in: query
          type: integer
          format: int64
          description: "Unique identifier for the user."
        - name: funds
          in: query
          type: number
          format: double
          description: "Amount to add to the user's funds."
      responses:
        200:
          description: "Updated user details with added funds."
          schema:
            $ref: "#/definitions/User"
        400:
          description: "Invalid user ID or amount."
        404:
          description: "User not found."
        500:
          description: "Internal Server Error."